FUSE - Change win_error err type to long (NTSTATUS)

+++ dokan_fuse/include/fuse_win.h
@@ -30,8 +30,8 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
-int ntstatus_error_to_errno(int win_res);
-int errno_to_ntstatus_error(int err);
+int ntstatus_error_to_errno(long win_res);
+long errno_to_ntstatus_error(int err);
 
 //This stuff is useful only on Windows in MSVC
 #ifdef _MSC_VER
