+++ dokan_fuse/include/fusemain.h
@@ -57,9 +57,9 @@ class win_error
 public:
 	win_error(int _err): err(errno_to_ntstatus_error(_err)) {}
 	win_error(int _err, bool): err(_err) {}
-	operator int() const { return err; }
+	operator long() const { return err; }
 private:
-	int err;
+	long err;
 };
 /*
 	FUSE filesystem context
